108991:kernel-modules-nvidia-un-def-331.20-alt1.199681.1 +11 more
- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-omnibook-un-def-20110911-alt1.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-virtualbox-un-def-4.2.18-alt1.199681.1 Evgeny Sinelnikov (Sin) <sin@altlinux.org>

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-bcmwl-un-def-6.30.223.30-alt1.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-image-un-def-3.12.1-alt1 Anton V. Boyarshinov <boyarsh@altlinux.ru>

- 3.12.1

kernel-modules-ipt_netflow-un-def-1.8-alt3.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-vhba-un-def-20130607-alt1.199681.1 Nazarov Denis (nenderus) <nenderus@altlinux.org>

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-bbswitch-un-def-0.7-alt1.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-r8168-un-def-8.037.00-alt1.199681.1 Nazarov Denis (nenderus) <nenderus@altlinux.org>

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-alx-un-def-20130405-alt1.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-emlog-un-def-0.51-alt3.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

kernel-modules-lsadrv-un-def-2.0.1-alt3.199681.1

- Build for kernel-image-un-def-3.12.1-alt1.

+++ k/kernel-modules-nvidia-un-def/kernel-modules-nvidia.spec
@@ -4,7 +4,7 @@
 %define nvIF_ver_lteq() %if "%(rpmvercmp '%2' '%1')" >= "0"
 
 %define module_name	nvidia
-%define module_version	319.72
+%define module_version	331.20
 %define module_release	alt1
 %define flavour		un-def
 
@@ -115,8 +115,10 @@ do
     tar -jxvf %_usrsrc/kernel/sources/kernel-source-%module_name-$sffx.tar.bz2
 
     pushd kernel-source-%module_name-$sffx
-    rm -f makefile Makefile
-    ln -s Makefile.kbuild Makefile
+    if [ -f Makefile.kbuild ] ; then
+	rm -f makefile Makefile
+	ln -s Makefile.kbuild Makefile
+    fi
     popd
 done
 
@@ -187,6 +189,9 @@ fi
 %changelog
 * %(date "+%%a %%b %%d %%Y") %{?package_signer:%package_signer}%{!?package_signer:%packager} %version-%release
 - Build for kernel-image-%flavour-%kversion-%krelease.
+
+* Tue Nov 19 2013 Sergey V Turchin <zerg at altlinux dot org> 331.20-alt1..
+- new release (331.20)
 
 * Thu Nov 07 2013 Sergey V Turchin <zerg at altlinux dot org> 319.72-alt1..
 - new releases (319.72, 304.116, 173.14.38)
